fix tests, add joi validation
