--- apps/server/Notes.md
@@ -1,3 +1,0 @@
-## Updating package:
-- Change version number in accordance with semver
-- run ``yarn publish --access=public``
