+++ apps/server/package.json
@@ -9,14 +9,12 @@
     "dev": "nodemon",
     "test": "jest --forceExit --detectOpenHandles --verbose",
     "test:watch": "jest --watch --forceExit --detectOpenHandles --verbose",
-    "coverage": "jest --forceExit --detectOpenHandles --coverage --verbose",
-    "test:json-server": "json-server --watch ./test/db.json --port 8010"
+    "coverage": "jest --forceExit --detectOpenHandles --coverage --verbose"
   },
   "dependencies": {
     "axios": "^1.2.1",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
-    "execa": "^6.1.0",
     "express": "^4.18.2",
     "joi": "^17.7.0",
     "json-server": "^0.17.1"
@@ -26,6 +24,7 @@
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.15",
     "@types/jest": "^29.2.4",
+    "@types/json-server": "^0.14.4",
     "@types/morgan": "^1.9.3",
     "@types/node": "^18.11.17",
     "@types/supertest": "^2.0.12",
