+++ apps/server/src/app.ts
@@ -4,10 +4,11 @@ import morgan from 'morgan';
 
 import apiRoutes from './routes';
 import { loggingType } from './configs';
+import { corsConfig } from './configs/cors';
 
 export const app = express();
 
-app.use(cors());
+app.use(cors(corsConfig));
 app.use(morgan(loggingType));
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
