+++ apps/server/src/configs/cors.ts
@@ -1,6 +1,7 @@
+import { CorsOptions } from 'cors';
 import { frontendUrl } from './index';
 
-export const corsConfig = {
+export const corsConfig: CorsOptions = {
   origin: [frontendUrl, 'http://localhost:3000'],
   // credentials: true,
   // methods: ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'], //notice OPTIONS not needed for preflight req to pass
