+++ apps/server/src/configs/index.ts
@@ -9,10 +9,12 @@ export const {
 
 const { FRONTEND_URL_DEV, FRONTEND_URL_PROD, DB_BASE_URL } = process.env;
 
-export const frontendUrl = NODE_ENV === 'production' ? FRONTEND_URL_PROD : FRONTEND_URL_DEV;
+const dbPort = NODE_ENV === 'test' ? DB_PORT_TEST : DB_PORT;
+
+export const frontendUrl = NODE_ENV === 'production' ? FRONTEND_URL_PROD as string : FRONTEND_URL_DEV as string;
 
 export const loggingType = NODE_ENV === 'production' ? 'combined' : 'dev';
 
-const dbPort = NODE_ENV === 'test' ? DB_PORT_TEST : DB_PORT;
-
 export const dbUrl = `${DB_BASE_URL}${dbPort}`;
+
+// export * from './cors'; // cannot access frontendUrl before init
