+++ apps/server/src/controllers/question.controller.ts
@@ -2,8 +2,7 @@ import axios from 'axios';
 import { Request, Response } from 'express';
 
 import { dbUrl } from '../configs';
-console.log('🚀  file: questions.controller.ts:5  dbUrl', dbUrl)
-import { questionModel } from '../models/question.validation';
+import { questionModel } from './question.validation';
 import { getErrorMessage } from '../utils';
 
 // @desc get all questions
@@ -23,19 +22,14 @@ export const getQuestions = async (req: Request, res: Response) => {
 // @access Private
 export const createQuestion = async (req: Request, res: Response) => {
   const { error, value } = questionModel.validate(req.body);
-  console.log('🚀  file: questions.controller.ts:25  value', value)
   if (error) {
     const { message } = error.details[0];
-    console.log('🚀  file: questions.controller.ts:28  message', message)
     return res.status(400).send(message);
   }
-  console.log('past erorr checkkkkkkkkkkkkkkkkkkkkkkkkkkkkk')
   try {
     const { data: newQuestion } = await axios.post(`${dbUrl}/questions`, { ...value });
-    console.log('🚀  file: questions.controller.ts:33  newQuestion', newQuestion)
     return res.status(201).send(newQuestion);
   } catch (error) {
-    console.log('🚀  file: questions.controller.ts:37  error', error)
     return res.status(500).send(getErrorMessage(error));
   }
 };
