+++ apps/server/tests/db.copy.json
@@ -87,7 +87,34 @@
           "score": 0.75
         }
       ]
+    },
+    {
+      "id": 5,
+      "text": "What is your favorite color?",
+      "options": [
+        {
+          "text": "Red",
+          "score": 0.5
+        },
+        {
+          "text": "Blue",
+          "score": 0.5
+        }
+      ]
+    },
+    {
+      "id": 6,
+      "text": "What is your favorite color?",
+      "options": [
+        {
+          "text": "Red",
+          "score": 0.5
+        },
+        {
+          "text": "Blue",
+          "score": 0.5
+        }
+      ]
     }
-    
   ]
 }
